python version up to 3.10.7, changed project structure, extracted repository, added match instead of if elif
