+++ .env
@@ -1,0 +1,1 @@
+POSTGRES_DB_URI=postgresql://postgres:postgres@localhost:5432/postgres
