+++ alembic/env.py
@@ -4,7 +4,6 @@ from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 
 from alembic import context
-import cv_creator.models.db_models
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
@@ -17,7 +16,7 @@ if config.config_file_name is not None:
 
 # add your model's MetaData object here
 # for 'autogenerate' support
-from cv_creator.database import Base
+from cv_creator.storage.postgres.database import Base
 target_metadata = Base.metadata
 # target_metadata = None
