+++ pyproject.toml
@@ -5,13 +5,14 @@ description = ""
 authors = ["Sebastian Przesmycki <temp_mail>"]
 
 [tool.poetry.dependencies]
-python = "^3.8.9"
+python = "^3.10.7"
 Flask = "^2.2.2"
 SQLAlchemy = "^1.4.40"
 marshmallow-sqlalchemy = "^0.28.1"
 alembic = "^1.8.1"
-psycopg2-binary = "^2.9.3"
+psycopg2-binary = "^2.9.4"
 marshmallow-dataclass = "^8.5.8"
+python-dotenv = "^0.21.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.1.2"
